Update SAAS service
The saas service is enabled by default now.
Remote HMAC auth checks now require a timestamp which is greater than 10 minutes ago. Should probably also do a check so its not too far in the future but is fine for now.
The saas setup process can be skipped by setting a BREAKFAST_PRE_SETUP env variable of a username to configure. A random password is generated so an admin or remote process will be needed to make the account accessible if the service is setup in this way.
Twitch auth provider can be setup with environment variables at launch.