Events now store an initiator and created viewers table
Events can store an initiator which will link a viewer who triggered the event.
Viewers table has also been created as an auth table. When an event calls for a viewer ID, if that viewer does not exist, then a viewer will be created with the external auth configured to the account.
Viewers also have an inventory column so can gamify things quickly. Though will need some controlled APIs to manage things.