Started moving bits into the breakfast package
Moving a lot of pieces into a library to help with templates and
creating offline editors and such. Things have been moved to its own
package and now just needs to convert the package styles to local
instead of tailwind.
Ideally the editor itself its made a part of the package so that it can
be dropped in anywhere and sync/save however the wrapping environment
sees fit.