Updated io defined editor to be able to handle lists
DefinedEditor can be configured to create a list of things from a given
input or array of inputs. If a single input is provided then the values
will most often be an array of strings. If an array of inputs is
provided then the values will be an array of objects with the usual
key:value structure like a normal Defined editor would give.