Changed twitch eventsub flow and minor improvements
Twitch eventsub would delete a record if it tried to subscribe but a
similar subscription already existed. I believe this is what causes #2
but I dont think this will fully resolve the issue. Probably need to
setup a scehule to check the active vs stored subscriptions in case
anything dies or needs to be resubscribed.
Have also improved docker image with some CA certs that might be a
problem (or not) and added env variable config for admin pass.