Server side of local overlays done and format
The server can now render and deliver local overlays that in theory
could be saved as a single html file and work as is using search params
for configuration. Still need to create the admin UI that manages and
lets users tweak configs and such.
Ran format in package project.